+++ poolparty/src/poolparty/orf_ops/__init__.py
@@ -1,6 +1,8 @@
 """ORF operations for poolparty."""
 
 from .annotate_orf import annotate_orf
+from .deletion_scan_orf import deletion_scan_orf
+from .insertion_scan_orf import insertion_scan_orf
 from .mutagenize_orf import MutagenizeOrfOp, mutagenize_orf
 from .reverse_translate import ReverseTranslateOp, reverse_translate
 from .stylize_orf import StylizeOrfOp, stylize_orf
@@ -8,6 +10,8 @@ from .translate import TranslateOp, translate
 
 __all__ = [
     "annotate_orf",
+    "deletion_scan_orf",
+    "insertion_scan_orf",
     "mutagenize_orf",
     "MutagenizeOrfOp",
     "reverse_translate",